pyproject.toml
[project]
name = "data-ingestion-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.55.0",
    "google-genai>=1.68.0",
    "pymongo[srv]>=4.16.0",
    "numpy>=2.4.3",
    "dotenv>=0.9.9",
]
